fonts, about page

+++ components/NewsItemFeatured.js
@@ -5,7 +5,7 @@ import CategoryLabel from './CategoryLabel'
 
 export default function NewsItemFeatured( {nw} ) {
 
-    const options = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
+    const options = { weekday: 'short', year: 'numeric', month: 'long', day: 'numeric' };
     
     return (
